MT#55283 fix naughty bit field bug
(cherry picked from commit 25a28db9ec9fa097c14f46b463d6e23f47b381ec)
(cherry picked from commit 0bdf058fa126d5f53dc6cc627cc673b0b6817814)

+++ lib/auxlib.h
@@ -647,7 +647,7 @@ INLINE void bf_copy(atomic64 *u, const uint64_t f,
 }
 /* works for multiple flags */
 INLINE void bf_copy_same(atomic64 *u, const atomic64 *s, const uint64_t g) {
-	unsigned int old, set, clear;
+	uint64_t old, set, clear;
 	old = atomic64_get(s);
 	set = old & g;
 	clear = ~old & g;
